Revert "chore: refactored <FilterHeader />"
This reverts commit 74ffe3cffc580977e533cb4ed35fcbc9882f397c.
